udpates file structure
